Add a condition on delete empty dir in backups

+++ scripts/perms.ksh
@@ -169,7 +169,8 @@ chown -R admin.www-data /home/admin/wwwroot/dolibarr_documents/sellyoursaas_loca
 chmod a+rwx $pathtospamdir; chmod a+rw $pathtospamdir/*
 chown admin.www-data $pathtospamdir/*
 
-
-echo "Nettoyage des repertoires de backup vides sous $backupdir/osu*"
-find $backupdir/osu*/ -type d -empty -ls -delete > /var/log/find_delete_empty_dir.log 2>&1
-
+if [ "x$instanceserver" != "x0" -a "x$instanceserver" != "x" ]; then
+	IFS=$(echo -en "\n\b")
+	echo "We are on a deployment server, so we try to delete empty dirs into backup directory under $backupdir/osu*"
+	find $backupdir/osu*/ -type d -empty -ls -delete > /var/log/find_delete_empty_dir.log 2>&1
+fi
